NOT READY FOR MERGE YET
- cvs upload test succeded 
- still need to test different upload options: dry run, product update, no stock update

+++ src/main/java/com/nasnav/commons/model/dataimport/ProductImportCsvRowData.java
@@ -12,6 +12,6 @@ public class ProductImportCsvRowData {
 	protected String barcode;
 	protected String category;
 	protected String brand;
-	protected Long quantity;
+	protected Integer quantity;
 	protected BigDecimal price;
 }
